requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-filter==25.1
drf-spectacular==0.27.2
celery==5.5.3
redis==6.4.0
reportlab==4.2.5
python-decouple==3.8
dj-database-url==2.2.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
python-memcached==1.62
django-celery-beat==2.8.0
django-celery-results==2.5.1

# Testing
factory-boy==3.3.1
pytest==8.4.1
pytest-django==4.9.0
pytest-mock==3.14.0
pytest-cov==6.0.0
